Refactor viz builtins module
