+++ app/gui2/env.d.ts
@@ -7,3 +7,9 @@ declare const PROJECT_MANAGER_URL: string
 interface ImportMeta {
   vitest: typeof import('vitest') | undefined
 }
+
+declare module 'builtins' {
+  export const VisualizationContainer: typeof import('@/components/VisualizationContainer.vue').default
+  export const useVisualizationConfig: typeof import('@/providers/visualizationConfig').useVisualizationConfig
+  export const defineKeybinds: typeof import('@/util/shortcuts').defineKeybinds
+}
