--- app/gui2/src/providers/useVisualizationConfig.ts
@@ -1,1 +1,0 @@
-export { useVisualizationConfig } from '@/providers/visualizationConfig.ts'
